% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SABARSI.R
\name{detect_sig}
\alias{detect_sig}
\title{Calculate the pvalues and false discovery rates (FDRs) for a background-removed spectrum}
\usage{
detect_sig(xr)
}
\arguments{
\item{xr}{A p-length vector that represents a background-removed spectrum with p frequency channels.}
}
\value{
A list containing a vector of pvalues, \code{pvals}, and a vector of FDRs, \code{fdrs}.
}
\description{
Calculate the pvalues and false discovery rates (FDRs) for a background-removed spectrum
}
