% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SABARSI.R
\name{shift_match}
\alias{shift_match}
\title{Match signals from two experiments.
For each signal in the first experiment, \code{shift.match} function finds the best matched signal in the second experiment. This function takes the potential frequency shifts into consideration for similarity measurement.}
\usage{
shift_match(xra, xrb, ta, tb)
}
\arguments{
\item{xra}{A p*n data matrix of background-removed spectra in the first experiment.}

\item{xrb}{A p*n data matrix of background-removed spectra in the second experiment.}

\item{ta}{A vector of time indices of signals in the first experiment.}

\item{tb}{A vector of time indices of signals in the second experiment.}
}
\value{
A list containing the time indices of signals in the first experiment, \code{ta}, and the time indices of corresponding time indices in the second experiment, as well as the correlation of each match pairs, \code{corra}.
}
\description{
Match signals from two experiments.
For each signal in the first experiment, \code{shift.match} function finds the best matched signal in the second experiment. This function takes the potential frequency shifts into consideration for similarity measurement.
}
